app/requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi @ file:///wheels/fastapi-0.110.1-py3-none-any.whl#sha256=5df913203c482f820d31f48e635e022f8cbfe7350e4830ef05a3163925b1addc
greenlet==3.0.3
h11==0.14.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
passlib==1.7.4
pydantic @ file:///wheels/pydantic-2.7.0-py3-none-any.whl#sha256=9dee74a271705f14f9a1567671d144a851c675b072736f0a7b2608fd9e495352
pydantic_core==2.18.1
PyJWT==2.8.0
python-multipart==0.0.9
setuptools==69.5.1
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typing_extensions==4.12.2
uvicorn @ file:///wheels/uvicorn-0.29.0-py3-none-any.whl#sha256=2c2aac7ff4f4365c206fd773a39bf4ebd1047c238f8b8268ad996829323473de
wheel==0.43.0
